Добавлен функционал проверки столкновения выстарела с препятствием:
- проверяется позиция пули на каждом такте
- при столкновении корутина пули удаляется
